FIR find usage: create separate KtFirSyntheticPropertyAccessorReference class
The existing code uses FirSimpleNameReference class to handle the
references to the Java synthetic property. It causes an issue in
`processReference(..)` method of MethodTextOccurrenceProcessor when the
property is used by read/write. Since read/write has two results of the
reference resolution i.e., getter for read and setter for write,
`ref.solve()` in the method returns `null` because it uses `resolve()`
method in PsiPolyVariantReferenceBase that returns null when it has
multiple results.

In the case of FE1.0, `ref.isReferenceTo(method)` in the method is true
when `method` is the getter and `ref` is the reference to the synthetic
property. On the other hand, `ref.isReferenceTo(method)` in the method
is false for FIR because FIR does not use "reference to Java synthetic
property" but it uses FirSimpleNameReference. Creating a separate class
for the reference to the Java synthetic property i.e.,
KtFirSyntheticPropertyAccessorReference class fixes the issue.